Added non-alloc helper methods
During queries, a lot of extra lists can be generated. Several new versions of methods that take an output list have been added. Internally the PopulationSample now uses these methods with static lists that are cleared with each use to help cut down on the number of allocations and reduce pressure on the GC.